Setup nginx on app3 for media serving on the streamwebs websites
The media folder on Django requires a webserver to serve them and this basically
implements it for the streamwebs production and staging websites. It sets it up
so that /<app name>/<whatever dir you need> is redirected properly from HAproxy.

So HAProxy knows that when it connects to this server, it needs to access
/streamwebs-production/media when a user requests /media.